backend/requirements.txt
fastapi
uvicorn
influxdb-client
paho-mqtt
